+++ example/app.dart
@@ -12,20 +12,45 @@ import 'package:polymer_app_router/polymer_app_router.dart';
 
 @PolymerRegister("polymer-app")
 class PolymerApp extends PolymerElement {
-  PolymerApp.created() : super.created();
+  PolymerApp.created() : super.created() {
+    homeRouteElem = createElement("home");
+    oneRouteElemA = createElement("one A");
+    oneRouteElemAA = createElement("one AA");
+    twoRouteElem = createElement("two");
+    oneRouteElemB = createElement("one B");
+
+    pages = [
+      new Page("home", "", element: homeRouteElem),
+      new Page("one", "one", isAbstract: true, redirectTo: "oneA"),
+      new Page("oneA", "a", element: oneRouteElemA, parent: "one"),
+      new Page("oneAA", "aa", element: oneRouteElemA, parent: "one"),
+      new Page("oneB", "b", element: oneRouteElemA, parent: "one"),
+      new Page("two", "two", element: twoRouteElem)
+    ];
+  }
+
+  PolymerAppRoute homeRouteElem;
+  PolymerAppRoute oneRouteElemA;
+  PolymerAppRoute oneRouteElemAA;
+  PolymerAppRoute oneRouteElemB;
+  PolymerAppRoute twoRouteElem;
 
   PolymerAppRoute createElement(String text) {
-    PolymerAppRoute route = (document.createElement("polymer-app-route") as PolymerAppRoute);
+    PolymerAppRoute route =
+        (document.createElement("polymer-app-route") as PolymerAppRoute);
     route.innerHtml = text;
     return route;
   }
 
+  List<Page> _pages;
+
+  set pages(List<Page> values) {
+    _pages = values;
+    notifyPath("pages", values);
+  }
+
   @property
-  List<Page> get pages =>  [
-    new Page("home", "", createElement("home")),
-    new Page("one", "/one", createElement("one")),
-    new Page("two", "/two", createElement("two"))
-  ];
+  List<Page> get pages => _pages;
 
   String _selectedMenu;