+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
 
-## 0.0.2
+## 0.2.0
+- hierarchical routes now supported (a/b/c ...)
+- ***Breaking Changes***: the element in the page class is now optional
+
+## 0.1.2
 - Remove polymer_elements (iron-pages)
 - Add Readme
 
 
-## 0.0.1
+## 0.1.1
 
 - Initial version